Added GeoSim and STS-EPR generative models
Added two individual generative models (GeoSim and STS-EPR) for human mobility, that model the social dimension togheter with the spatial and temporal dimensions of mobility.
The relative tests verify the correctness of the models, of the pseudo-random functions used in the models, the behavior with incorrect inputs, and of the choices made by the synthetic agents.