Managing SSL certificate exception
The error appears when trying to load the 'taxi_san_francisco' dataset. A SSLError is raised when checking the certificate with the 'requests' library. 

I added a new function to make it more readable the option to bypass the error if it appears. Avoiding the certificate check is not the safest solution, thus before proceeding I printed the hash that will be used to check the authenticity of the file with the 'pooch' library, if it is provided. If not provided, the known hash presented will be "None", making the user aware of the fact the file may not be the same they intended to download.